+++ API/Github/issues.py
@@ -1,4 +1,4 @@
-"""API interface Module for Git Issues calls"""
+"""API interface Module for Github Issues calls"""
 import requests
 
 from API.Github.github_base_calls import GithubBaseCalls
@@ -8,13 +8,13 @@ from API.Github.github_base_calls import GithubBaseCalls
 ENDPOINT = "issues"
 
 class Issues():
-    """Class for Git Issues API functions"""
+    """Class for Github Issues API functions"""
 
     def GetIssues(isJpl:False, owner:str, repoName:str, logging:bool=True) -> requests.Response:
         """Function to call the issues endpoint"""
 
         if logging:
-            print(f'Getting data from Git "/{ENDPOINT}" endpoint...')
+            print(f'Getting data from Github "/{ENDPOINT}" endpoint...')
 
         return GithubBaseCalls.Get(
             endpoint = f'/repos/{owner}/{repoName}/{ENDPOINT}',