+++ data_updater/pypi.py
@@ -6,6 +6,6 @@ class PyPi():
     """Class for PyPi related methods"""
 
     def GetPyPiReleaseLink(environment:str) -> str:
-        """Function to get the release link of the git repository published to PyPi"""
+        """Function to get the release link of the Github repository published to PyPi"""
 
         return "Under development"